fix: fix destroy() with withTrashed, remove withTrashed from index()

+++ app/Http/Controllers/Admin/AdminRoomTypeController.php
@@ -18,8 +18,7 @@ class AdminRoomTypeController extends Controller
     {
         try {
             // Sử dụng withTrashed() để kéo lại toàn bộ phòng cũ/mới bị ẩn do dính cột deleted_at
-            $roomTypes = RoomType::withTrashed()
-                ->with('images')
+            $roomTypes = RoomType::with('images')
                 ->orderBy('id', 'desc')
                 ->get()
                 ->map(function ($roomType) {
@@ -301,12 +300,15 @@ class AdminRoomTypeController extends Controller
      */
     public function destroy($id)
     {
-        $roomType = RoomType::findOrFail($id);
+        $roomType = RoomType::withTrashed()->findOrFail($id);
+
+        if ($roomType->trashed()) {
+            return response()->json(['message' => 'Loại phòng đã bị xóa trước đó'], 400);
+        }
 
         DB::beginTransaction();
 
         try {
-            // Không xóa ảnh thật trong ổ cứng khi xóa mềm để có thể dùng hàm restore khôi phục lại nguyên vẹn
             $roomType->delete();
             DB::commit();
